# src/main/resources/api-specification.yml
openapi: 3.1.0
info:
  title: Modified Method Of CMLR
  version: 1.0.0

servers:
  - url: http://localhost:3000
    description: Local server

tags:
  - name: MLR
    description: Multivariate Linear Regression operations
  - name: Generator
    description: Generator operations

paths:
  /multivariate-linear-regression/modified-method-of-cmlr/normally-distributed-random-numbers:
    post:
      summary: Executes modified method of CMLR with given number of runs with random variables as normally distributed random numbers
      operationId: solveWithModifiedMethodOfCmlrWithRandomVariablesAsNormallyDistributedRandomNumbers
      tags:
        - MLR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NormallyDistributedRandomNumbersMmcmlrRequestBody"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegressionSimulationStatisticsResponse"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestResponse"

  /multivariate-linear-regression/modified-method-of-cmlr/uniformly-distributed-random-numbers:
    post:
      summary: Executes modified method of CMLR with given number of runs with random variables as uniformly distributed random numbers
      operationId: solveWithModifiedMethodOfCmlrWithRandomVariablesAsUniformlyDistributedRandomNumbers
      tags:
        - MLR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UniformlyDistributedRandomNumbersMmcmlrRequestBody"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegressionSimulationStatisticsResponse"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestResponse"

  /random-numbers/uniformly-distribution/int/matrix:
    post:
      summary: Generate matrix of uniformly distributed int random numbers
      operationId: generateMatrixOfUniformlyDistributedIntRandomNumbers
      tags:
        - Generator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntRandomNumbersRequestBody"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: array
                  items:
                    type: number
                    format: double
                example: [[1, 5], [3, 8]]
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestResponse"

  /random-numbers/uniformly-distribution/double/matrix:
    post:
      summary: Generate matrix of uniformly distributed double random numbers
      operationId: generateMatrixOfUniformlyDistributedDoubleRandomNumbers
      tags:
        - Generator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoubleRandomNumbersRequestBody"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: array
                  items:
                    type: number
                    format: double
                example: [[1.3, 5.1], [3.6, 8.4]]
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestResponse"

components:
  schemas:
    UniformlyDistributedRandomNumbersMmcmlrRequestBody:
      type: object
      description: Request body for running multiple times modified method of CMLR with random variables as uniformly distributed random numbers
      required:
        - repetitionsNumberOfActiveExperiments
        - numberOfValidationSequences
        - numberOfRuns
        - correctCoefficients
        - independentVariables
        - range
      properties:
        repetitionsNumberOfActiveExperiments:
          type: integer
          minimum: 1
          maximum: 10
          description: Repetitions Number of Active Experiments
          example: 2
        numberOfValidationSequences:
          type: integer
          minimum: 1
          maximum: 10
          description: Number of validation sequences
          example: 2
        numberOfRuns:
          type: integer
          minimum: 1
          maximum: 10000
          description: Number of runs
          example: 1000
        correctCoefficients:
          type: array
          items:
            type: number
            format: double
          minItems: 1
          maxItems: 30
          description: Correct coefficients array
          example: [1.0, 0, 3.0, 0, 6.0]
        independentVariables:
          type: array
          items:
            type: array
            items:
              type: number
              format: double
            minItems: 1
            maxItems: 30
          minItems: 1
          maxItems: 300
          description: Independent variables matrix
          example: [
              [4, 1, 1, 1],
              [4, 4, 1, 1],
              [1, 4, 4, 1],
              [1, 1, 4, 4],
              [4, 4, 4, 1],
              [1, 4, 4, 4],
              [1, 1, 4, 4],
              [1, 1, 1, 4],
              [4, 4, 4, 4],
              [4, 1, 1, 1]
            ]
        range:
          $ref: "#/components/schemas/DoubleRange"

    NormallyDistributedRandomNumbersMmcmlrRequestBody:
      type: object
      description: Request body for running multiple times modified method of CMLR with random variable as normally distributed random numbers
      required:
        - repetitionsNumberOfActiveExperiments
        - numberOfValidationSequences
        - numberOfRuns
        - correctCoefficients
        - independentVariables
        - mean
        - stdDev
      properties:
        repetitionsNumberOfActiveExperiments:
          type: integer
          minimum: 1
          maximum: 10
          description: Repetitions Number of Active Experiments
          example: 2
        numberOfValidationSequences:
          type: integer
          minimum: 1
          maximum: 10
          description: Number of validation sequences
          example: 2
        numberOfRuns:
          type: integer
          minimum: 1
          maximum: 10000
          description: Number of runs
          example: 1000
        correctCoefficients:
          type: array
          items:
            type: number
            format: double
          minItems: 1
          maxItems: 30
          description: Correct coefficients array
          example: [1.0, 0, 3.0, 0, 6.0]
        independentVariables:
          type: array
          items:
            type: array
            items:
              type: number
              format: double
            minItems: 1
            maxItems: 30
          minItems: 1
          maxItems: 300
          description: Independent variables matrix
          example: [
              [4, 1, 1, 1],
              [4, 4, 1, 1],
              [1, 4, 4, 1],
              [1, 1, 4, 4],
              [4, 4, 4, 1],
              [1, 4, 4, 4],
              [1, 1, 4, 4],
              [1, 1, 1, 4],
              [4, 4, 4, 4],
              [4, 1, 1, 1]
            ]
        mean:
          type: number
          description: Mathematical expectation
          example: 0
        stdDev:
          type: number
          description: Standard deviation
          example: 1

    IntRandomNumbersRequestBody:
      type: object
      description: Request body for generating matrix of random integer numbers
      required:
        - rows
        - columns
        - range
      properties:
        rows:
          type: integer
          minimum: 1
          description: Number of matrix rows
          example: 2
        columns:
          type: integer
          minimum: 1
          description: Number of matrix columns
          example: 2
        range:
          $ref: "#/components/schemas/IntRange"

    DoubleRandomNumbersRequestBody:
      type: object
      description: Request body for generating matrix of random double numbers
      required:
        - rows
        - columns
        - range
      properties:
        rows:
          type: integer
          minimum: 1
          description: Number of matrix rows
          example: 2
        columns:
          type: integer
          minimum: 1
          description: Number of matrix columns
          example: 2
        range:
          $ref: "#/components/schemas/DoubleRange"

    RegressionSimulationStatisticsResponse:
      type: object
      description: Response for running modified method of CMLR multiple times
      properties:
        percentageOfCorrectModels:
          type: number
          description: Percentage of correct models comparing to total number of runs
          example: 70.8
        percentageOfIncorrectModelsWithOneIncorrectZero:
          type: number
          description: Percentage of incorrect models with 1 incorrect zero comparing to total number of runs
          example: 20.2
        percentageOfIncorrectModelsWithTwoPlusIncorrectZeros:
          type: number
          description: Percentage of incorrect models with 2+ incorrect zeros comparing to total number of runs
          example: 9
        meanComparisonMeasureValueForCorrectModels:
          type: number
          description: Mean comparison measure value for correct models. If no correct models - returns -1
          example: 0.01324
        minComparisonMeasureValueForCorrectModels:
          type: number
          description: Min comparison measure value for correct models. If no correct models - returns -1
          example: 0.00357
        maxComparisonMeasureValueForCorrectModels:
          type: number
          description: Max comparison measure value for correct models. If no correct models - returns -1
          example: 0.43246

    BadRequestResponse:
      type: object
      description: Response for 400 HTTP status
      properties:
        globalErrors:
          type: array
          items:
            type: string
            example: Errors rows number should be equal to independent variables rows number
          description: Errors based on multiple fields
        fieldErrors:
          type: array
          items:
            type: string
            example: Independent variables matrix is singular
          description: Errors based on specific field

    IntRange:
      type: object
      description: Range of integer numbers
      properties:
        start:
          type: integer
          description: Start of range inclusively
          example: 0
        end:
          type: integer
          description: End of range exclusively
          example: 10

    DoubleRange:
      type: object
      description: Range of double numbers
      properties:
        start:
          type: number
          format: double
          description: Start of range inclusively
          example: 0.5
        end:
          type: number
          format: double
          description: End of range exclusively
          example: 10.3